+++ src/parser/companions/types/FindFamiliar.mjs
@@ -195,7 +195,7 @@ export async function getFindFamiliarActivityData(activity, options) {
   const profiles = [];
 
   for (const familiar of mapInUse) {
-    const i = ddbCompendium?.index.find((i) => i.name === familiar.name && i.system.source.rules === rules);
+    const i = ddbCompendium?.index.find((i) => i.name === familiar.name && i.system?.source?.rules === rules);
     if (i) profiles.push({
       name: familiar.name,
       uuid: i.uuid,