+++ src/parser/enrichers/feat/_module.mjs
@@ -20,6 +20,7 @@ export { default as Healer } from "./Healer.mjs";
 export { default as HeavyArmorMaster } from "./HeavyArmorMaster.mjs";
 export { default as InspiringLeader } from "./InspiringLeader.mjs";
 export { default as MageSlayer } from "./MageSlayer.mjs";
+export { default as MetamagicAdept } from "./MetamagicAdept.mjs";
 export { default as Observant } from "./Observant.mjs";
 export { default as Piercer } from "./Piercer.mjs";
 export { default as Poisoner } from "./Poisoner.mjs";