+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Font of Magic generated/added sorcery point value to it as wel as the Sorcery Points feature.
 - Bard Feature: Blade Flourish corrections.
 - Item enhancements: Air Render, Siege Staff @eikka
+- Sorcerer Sorcery Points corrections with Metamagic Adept. @co6dm
 - DDB Muncher Screen migrated to Application V2.
 
 # 6.0.89